# packages/api-contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Koupreng E-Invitation API
  description: |
    REST API for the Koupreng E-Invitation platform, including authentication,
    users, invitations, templates, guests, RSVP, check-in, media, payments,
    organizations, subscriptions, notifications, reporting, administration,
    and related services. The current contract includes both legacy /api routes
    and versioned /api/v1 routes; the version below does not imply a global base path.
  version: v1
servers:
- url: http://localhost
  description: Generated server url
...
paths:
...
  /api/v1/ai/invitation/translate:
    post:
      tags:
      - AI Invitation Assistant
      summary: Translate invitation copy
      operationId: translate
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiInvitationDraftRequest"
            examples:
              bilingualWedding:
                summary: Generate fictional bilingual wedding copy
                value:
                  language: BILINGUAL
                  tone: warm and formal
                  eventType: WEDDING
                  coupleNames: Sokha and Pisey
                  hostName: The Sok and Lim families
                  venueName: Phnom Penh Demo Ballroom
                  eventDate: 2035-02-17
                  notes: Include a concise Khmer welcome and an English translation.
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseAiInvitationDraftResponse"
      security:
      - bearerAuth: []
  /api/v1/ai/invitation/timeline-suggestion:
    post:
      tags:
      - AI Invitation Assistant
      summary: Suggest an invitation timeline
      operationId: timelineSuggestion
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiInvitationDraftRequest"
            examples:
              bilingualWedding:
                summary: Generate fictional bilingual wedding copy
                value:
                  language: BILINGUAL
                  tone: warm and formal
                  eventType: WEDDING
                  coupleNames: Sokha and Pisey
                  hostName: The Sok and Lim families
                  venueName: Phnom Penh Demo Ballroom
                  eventDate: 2035-02-17
                  notes: Include a concise Khmer welcome and an English translation.
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseAiInvitationDraftResponse"
      security:
      - bearerAuth: []
  /api/v1/ai/invitation/story:
    post:
      tags:
      - AI Invitation Assistant
      summary: Generate an invitation story
      operationId: story
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiInvitationDraftRequest"
            examples:
              bilingualWedding:
                summary: Generate fictional bilingual wedding copy
                value:
                  language: BILINGUAL
                  tone: warm and formal
                  eventType: WEDDING
                  coupleNames: Sokha and Pisey
                  hostName: The Sok and Lim families
                  venueName: Phnom Penh Demo Ballroom
                  eventDate: 2035-02-17
                  notes: Include a concise Khmer welcome and an English translation.
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseAiInvitationDraftResponse"
      security:
      - bearerAuth: []
  /api/v1/ai/invitation/formal-text:
    post:
      tags:
      - AI Invitation Assistant
      summary: Generate formal invitation text
      operationId: formalText
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiInvitationDraftRequest"
            examples:
              bilingualWedding:
                summary: Generate fictional bilingual wedding copy
                value:
                  language: BILINGUAL
                  tone: warm and formal
                  eventType: WEDDING
                  coupleNames: Sokha and Pisey
                  hostName: The Sok and Lim families
                  venueName: Phnom Penh Demo Ballroom
                  eventDate: 2035-02-17
                  notes: Include a concise Khmer welcome and an English translation.
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseAiInvitationDraftResponse"
      security:
      - bearerAuth: []
  /api/v1/ai/invitation-copy:
    post:
      tags:
      - AI Invitation Assistant
      summary: Draft invitation copy
      operationId: draft
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiInvitationDraftRequest"
            examples:
              bilingualWedding:
                summary: Generate fictional bilingual wedding copy
                value:
                  language: BILINGUAL
                  tone: warm and formal
                  eventType: WEDDING
                  coupleNames: Sokha and Pisey
                  hostName: The Sok and Lim families
                  venueName: Phnom Penh Demo Ballroom
                  eventDate: 2035-02-17
                  notes: Include a concise Khmer welcome and an English translation.
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseAiInvitationDraftResponse"
      security:
      - bearerAuth: []
...
components:
  schemas:
...
    AiInvitationDraftRequest:
      type: object
      properties:
        language:
          type: string
          maxLength: 60
          minLength: 0
        tone:
          type: string
          maxLength: 60
          minLength: 0
        eventType:
          type: string
          maxLength: 120
          minLength: 0
        coupleNames:
          type: string
          maxLength: 240
          minLength: 0
        hostName:
          type: string
          maxLength: 120
          minLength: 0
        venueName:
          type: string
          maxLength: 240
          minLength: 0
        eventDate:
          type: string
          maxLength: 60
          minLength: 0
        notes:
          type: string
          maxLength: 5000
          minLength: 0
    AiInvitationDraftResponse:
      type: object
      properties:
        enabled:
          type: boolean
        provider:
          type: string
        source:
          type: string
        operation:
          type: string
        generatedText:
          type: string
        suggestions:
          type: array
          items:
            type: string
        warnings:
          type: array
          items:
            type: string
    ApiResponseAiInvitationDraftResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: "#/components/schemas/AiInvitationDraftResponse"
        timestamp:
          type: string
          format: date-time
...
  securitySchemes:
    bearerAuth:
      type: http
      description: Paste the accessToken returned by POST /api/auth/login. Scalar
        adds the Bearer prefix automatically; do not type it twice.
      scheme: bearer
      bearerFormat: JWT
    adminPaymentSecret:
      type: apiKey
      description: Internal Telegram payment reconciliation secret.
      name: X-ADMIN-PAYMENT-SECRET
      in: header